+++ TagsSearch/IndropV1TagsFinder.h
@@ -29,7 +29,7 @@ namespace TagsSearch
 
 	public:
 		explicit IndropV1TagsFinder(const std::vector<std::string> &fastq_filenames,
-				                    const boost::property_tree::ptree &spacer_config, const boost::property_tree::ptree &config,
+				                    const boost::property_tree::ptree &spacer_config, const boost::property_tree::ptree &processing_config,
 		                            const std::shared_ptr<ConcurrentGzWriter> &writer, bool save_stats, bool save_read_params);
 	};
 }